Setting Up Database, Fetching Data
This commit revitalizes the dashboard interface by uncommenting key sections of code and adding functionality to retrieve and display financial data. It also incorporates a new script, "seed", to the package.json file, providing a method for populating the database with sample data. These changes improve the dashboard's usability and support data testing.